+++ src/engine/kinematics/index.ts
@@ -5,4 +5,6 @@ export * from './poseDetector';
 export * from './occlusionHeuristics';
 export * from './rulesEngine';
 export * from './referenceDataLoader';
+export * from './dtwCalculator';
+